Preserve full 16.16 precision in CPU graphics matrices (#214)
Preserve every decoded signed 16.16 bit in CPU graphics matrices by
using Float64Array for loading, construction, copying, and
multiplication. A coefficient of `256 + 1/65536` now survives
cancellation against `-256`, producing homogeneous `w = 1/65536` instead
of zero. Existing Float32 vertex/rendering storage remains unchanged; no
new RSP fixed-point truncation rules are introduced.

Adds regressions for positive and negative precision boundaries, copies,
matrix products, and homogeneous-w cancellation through rendering
storage.

Validation: all 1,358 tests pass; lint and production build pass.

Fixes #194.

## Performance comparison

Comparison: `355caf55` (base) versus `277c5c2` (PR). Bun 1.3.14, macOS
arm64.

Eight paired runs per workload, in separate processes, alternating
base/head order. Runs execute serially. Each game sample boots a fresh
emulator, warms up for 120 VI retraces, then measures 1,800 retraces.
The existing benchmark harness was changed identically in both temporary
checkouts to pass `{ executeGraphics: true }` to
`createHeadlessEmulator`.

Positive time changes mean slower; negative means faster.

| Workload | Base median | PR median | Time change | Base / PR MAD |
|---|---:|---:|---:|---:|
| Super Mario 64 (USA) | 20.578 s | 20.312 s | -1.30% | 0.188 / 0.121 s
|
| Dark Rift (USA) | 4.224 s | 4.181 s | -1.02% | 0.037 / 0.057 s |
| Diddy Kong Racing (USA, Rev 1) | 13.004 s | 13.053 s | +0.38% | 0.042
/ 0.065 s |

Each base/PR game pair has identical executed cycles, retrace counts,
and final CPU-state fingerprint: True.

The graphics path uses NullRenderer: these results cover CPU and HLE
display-list work, not GPU rendering or browser frame rate. They cover
boot/intro sequences without controller input.

No consistent game-level slowdown was established in this session:
approximate 95% paired bootstrap intervals for time changes include zero
for all three games (Mario −1.43% to +0.62%, Dark Rift −3.20% to +1.62%,
DKR −0.56% to +1.45%). These are exploratory intervals from eight pairs,
not proof of equivalence.

Isolated matrix operations use 100,000 warm-up iterations followed by
500,000 measured iterations per operation per process; medians over
eight processes per version. Outputs contribute to a checksum. Fixture
matrices contain exactly representable values, avoiding divergent inputs
between versions.

| Operation | Base ns/op | PR ns/op | Time change |
|---|---:|---:|---:|
| load | 43.49 | 47.50 | +9.21% |
| copy | 12.30 | 16.80 | +36.61% |
| multiply | 30.26 | 28.07 | -7.25% |
| transform | 8.82 | 8.45 | -4.19% |
| loadMultiplyTransform | 74.96 | 70.81 | -5.54% |


Matrix loading and copying are measurably slower in isolation, while
multiplication and the combined load–multiply–transform workload are
faster in this Bun session. No consistent game-level slowdown was
established in these sampled boot/intro workloads.

+++ src/hle/gbi_microcode.js
@@ -124,7 +124,8 @@ export class GBIMicrocode {
       this.warn('Unusual matrix length', `${length}`);
     }
 
-    const elements = new Float32Array(16);
+    // All signed 16.16 values are exactly representable in float64.
+    const elements = new Float64Array(16);
     for (let i = 0; i < 4; ++i) {
       elements[4 * 0 + i] = (dv.getInt16(i * 8 + 0) << 16 | dv.getUint16(i * 8 + 0 + 32)) * recip;
       elements[4 * 1 + i] = (dv.getInt16(i * 8 + 2) << 16 | dv.getUint16(i * 8 + 2 + 32)) * recip;
